requirements.txt
streamlit>=1.40.0
google-genai
python-dotenv
Pillow